requirements.txt
# ---------------------------------------------------------------
# AI Safety Monitoring — finetune branch
# Python >= 3.11
#
# NOTE: Install PyTorch with CUDA *before* running pip install -r:
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cu124
# ---------------------------------------------------------------

# Deep Learning
torch>=2.2.2
torchvision>=0.17.2
timm==0.9.16

# YOLO (scene segmentation + human detection)
ultralytics>=8.3.252

# Image processing
opencv-python>=4.9.0.80
Pillow>=10.3.0
numpy>=1.26.4

# Data / training utilities
scikit-learn>=1.3.0
scipy>=1.11.0
matplotlib>=3.8.0
tqdm>=4.65.0

# Config
PyYAML>=6.0.1
